Move scrolling code into a dedicated scrolling module
