+++ packages/blitz-dom/src/lib.rs
@@ -52,6 +52,8 @@ mod layout;
 mod mutator;
 mod query_selector;
 mod resolve;
+/// Scrolling of nodes and the viewport, and scroll animations.
+mod scrolling;
 mod selection;
 /// Implementations that interact with servo's style engine
 mod stylo;
@@ -78,10 +80,7 @@ pub use crate::node::Widget;
 
 pub use blitz_traits::node_id::NodeId;
 pub use config::{DocumentConfig, StyleThreading};
-pub use document::{
-    BaseDocument, DocGuard, DocGuardMut, Document, PlainDocument, ScrollBehavior,
-    ScrollLogicalPosition,
-};
+pub use document::{BaseDocument, DocGuard, DocGuardMut, Document, PlainDocument};
 pub use markup5ever::{
     LocalName, Namespace, NamespaceStaticSet, Prefix, PrefixStaticSet, QualName, local_name,
     namespace_prefix, namespace_url, ns,
@@ -89,6 +88,7 @@ pub use markup5ever::{
 pub use mutator::DocumentMutator;
 pub use node::{Attribute, DocumentData, ElementData, Node, NodeData, TextNodeData};
 pub use parley::FontContext;
+pub use scrolling::{ScrollBehavior, ScrollLogicalPosition};
 pub use tree::NodeTree;
 
 /// Convert a Blitz [`NodeId`] into a [`taffy::NodeId`] (which wraps a `u64`).
